+++ CHANGELOG.md
@@ -64,6 +64,8 @@
 32. Added Start Screen Image Filtering Customization.
 33. MelonPreferences_Category DisplayName and IsHidden is now Modifiable during Runtime.
 34. MelonPreferences_Entry DisplayName, Description, IsHidden, and DontSaveDefault are now Modifiable during Runtime.
+35. Added MelonPreferences_Category.IsInlined for inlining the TomlTable of the Category.
+36. MelonPreferences_Entry.Description now Writes to File as a Comment.
 
 ---